redis final commit

+++ src/main/java/com/example/springbootredis/service/EmployeeService.java
@@ -1,9 +1,12 @@
 package com.example.springbootredis.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
 import org.springframework.stereotype.Service;
 
 import com.example.springbootredis.model.Employee;
@@ -15,24 +18,58 @@ public class EmployeeService {
 	@Autowired
 	private EmployeeRepository empRepository;
 	
-	public Optional<Employee> getEmployeeDetails(Integer empId) {
-		Optional<Employee> empDetails = empRepository.findById(empId);
-		if(empDetails != null) {
-			return empDetails;
+	@Cacheable(value = "empCache", key="#empId")
+	public Employee getEmployeeDetails(Integer empId) {
+		if(empRepository.findById(empId).isPresent())
+			return empRepository.findById(empId).get();
+		else {
+			return new Employee();
 		}
-		return null;
 	}
 	
+	@Cacheable(value = "allEmpCache", unless = "#result.size() == 0")
 	public List<Employee> getAllEmpDetails() {
 		return empRepository.findAll();
 	}
 	
-	public void deleteEmpDetails(Integer empId) {
-		empRepository.deleteById(empId);
-		
+	@Caching(
+			evict = {
+					@CacheEvict(value = "empCache", key = "#empId"),
+					@CacheEvict(value = "allEmpCache", allEntries = true)
+					}
+			)
+	@CacheEvict(value = "allEmpCache", allEntries = true)
+	public String deleteEmpDetails(Integer empId) {
+		if(empRepository.findById(empId).isPresent()) {
+			empRepository.deleteById(empId);
+			return "Deleted Successfully...";
+		} else {
+			return "No data with given info";
+		}
 	}
-
-	public void saveEmp(Employee emp) {
-		empRepository.save(emp);
+	
+	@Caching(
+			put = { @CachePut(value = "empCache", key = "#emp.empId") },
+			evict = { @CacheEvict(value = "allEmpCache", allEntries = true)}
+			)
+	public Employee saveEmp(Employee emp) {
+		return empRepository.save(emp);
+	}
+	
+	@Caching (
+		put = { @CachePut(value = "empCache", key = "#empId") },
+		evict = { @CacheEvict(value = "allEmpCache", allEntries = true)}
+	)
+	public Employee updateEmpDetails(Employee newEmp, Integer empId) {
+		if(empRepository.findById(empId).isPresent()) {
+			Employee emp = empRepository.findById(empId).get();
+			emp.setName(newEmp.getName());
+			emp.setPassword(newEmp.getPassword());
+			emp.setUsername(newEmp.getUsername());
+			return empRepository.save(emp);
+		} else {
+			return empRepository.save(newEmp);
+		}
+		
 	}
 }
